LAHAGORA_Scrapping_Task/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.11
filelock==3.8.0
google-play-scraper==1.2.2
greenlet==1.1.3
idna==3.4
kombu==5.2.4
langdetect==1.0.9
multipledispatch==0.6.0
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
phonenumbers==8.13.1
Pillow==9.3.0
platformdirs==2.5.2
prompt-toolkit==3.0.33
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
webencodings==0.5.1